Fix OTR query detection
New version of libotr changed the "missing plugin" message sent with
the OTR query (replacing http with https URLs). This broke the default
query workaround, resulting in IRC server complaining about invalid
commands.

To prevent such changes breaking znc-otr in the future, look only for
the actual OTR query (such as ?OTR? or ?OTR?v23?) in the message
instead of trying to match the message exactly.