Reassemble fragmented OTR messages
When using an XMPP to IRC gateway such as bitlbee, a single OTR
message can get broken into multiple parts due to IRC message length
limit. Buffer such parts until a whole OTR message is received, and
only then pass it to libotr.

Note that this is unrelated to fragmentation defined in OTR protocol,
which is handled by libotr.